refactor: Refactor env/ functions & header file
